change tests location
